Added Magic Multicoupler V1
This is a project that's been public for some time, but I finally got around to actually cleaning it up and releasing it. This is the next generation of coupling technologies; It's the original MCC, but better in every conceivable way. More efficient, capable of marking as many cars as you like for automatic coupling, reconfigurable, possessing a handbrake and coupling distance indicator, and having all the fancy modern visual effects to boot. Apologies that it took so long to upload, heh.